fix(edit): Fix creating additional loadables
Prior to this patch, for zpe-edit to open up your editor, the file had
to already exist.
However, it's common to want additional "custom commands", like
$name_push and $name_pull for loading with zpe-load.
This checks for a file already existing (to avoid readonly problems),
but also checks for $name being a registered project.
If the project exists, edit anyway.